add test cases to ClientMessageReaderTest.js from ClientMessageReaderTestRenewed.js and delete ClientMessageReaderTestRenewed.js.
