+++ test/unit/connection/ClientMessageReaderTest.js
@@ -20,6 +20,7 @@ const { expect } = require('chai');
 
 const { ClientMessageReader } = require('../../../lib/network/Connection');
 const cm = require('../../../lib/protocol/ClientMessage');
+const {Frame, ClientMessage} = require("../../../lib/protocol/ClientMessage");
 
 describe('ClientMessageReaderTest', function () {
     const HEADER_SIZE = cm.SIZE_OF_FRAME_LENGTH_AND_FLAGS;
@@ -30,6 +31,20 @@ describe('ClientMessageReaderTest', function () {
         const buffer = Buffer.allocUnsafe(cm.SIZE_OF_FRAME_LENGTH_AND_FLAGS + size);
         buffer.writeInt32LE(buffer.length, 0);
         buffer.writeUInt16LE(1 << 15 | 1 << 14 | 1 << 13, 4);
+        return buffer;
+    }
+
+    function createFrameWithRandomBytes(length) {
+        const buffer = Buffer.allocUnsafe(length);
+        for (let i = 0; i < length; i++) {
+            buffer.writeUInt8(Math.floor(Math.random() * 256), i);
+        }
+        return new Frame(buffer);
+    }
+
+    function writeToBuffer(message) {
+        const buffer = Buffer.allocUnsafe(message.getTotalLength());
+        message.writeTo(buffer);
         return buffer;
     }
 
@@ -97,5 +112,127 @@ describe('ClientMessageReaderTest', function () {
         reader.read();
 
         expect(reader.read()).to.be.equal(null);
+    });
+
+    it('testReadSingleFrameMessage', function() {
+        const frame = createFrameWithRandomBytes(42);
+        const message = ClientMessage.createForEncode();
+        message.addFrame(frame);
+        const buffer = writeToBuffer(message);
+        reader.append(buffer);
+
+        const messageRead = reader.read();
+        expect(messageRead).to.be.not.equal(null);
+        expect(messageRead.hasNextFrame()).to.be.true;
+
+        const frameRead = messageRead.nextFrame();
+        expect(frameRead.content).to.deep.equal(frame.content);
+
+        expect(messageRead.hasNextFrame()).to.be.false;
+    });
+
+    it('testReadMultiFrameMessage', function() {
+        const frame1 = createFrameWithRandomBytes(10);
+        const frame2 = createFrameWithRandomBytes(20);
+        const frame3 = createFrameWithRandomBytes(30);
+
+        const message = ClientMessage.createForEncode();
+        message.addFrame(frame1);
+        message.addFrame(frame2);
+        message.addFrame(frame3);
+
+        const buffer = writeToBuffer(message);
+        reader.append(buffer);
+
+        const messageRead = reader.read();
+        expect(messageRead).to.be.not.equal(null);
+        expect(messageRead.hasNextFrame()).to.be.true;
+        let frameRead = messageRead.nextFrame();
+        expect(frameRead.content).to.deep.equal(frame1.content);
+
+        expect(messageRead.hasNextFrame()).to.be.true;
+        frameRead = messageRead.nextFrame();
+        expect(frameRead.content).to.deep.equal(frame2.content);
+
+        expect(messageRead.hasNextFrame()).to.be.true;
+        frameRead = messageRead.nextFrame();
+        expect(frameRead.content).to.deep.equal(frame3.content);
+
+        expect(messageRead.hasNextFrame()).to.be.false;
+    });
+
+    it('testReadFramesInMultipleCallsToReadFrom', function() {
+        const frame = createFrameWithRandomBytes(1000);
+
+        const message = ClientMessage.createForEncode();
+        message.addFrame(frame);
+
+        const buffer = writeToBuffer(message);
+
+        const firstPartition = buffer.slice(0, 750);
+        const secondPartition = buffer.slice(750);
+
+        reader.append(firstPartition);
+        reader.append(secondPartition);
+
+        const messageRead = reader.read();
+        expect(messageRead.hasNextFrame()).to.be.true;
+
+        const frameRead = messageRead.nextFrame();
+        expect(frameRead.content).to.deep.equal(frame.content);
+
+        expect(messageRead.hasNextFrame()).to.be.false;
+    });
+
+    it('testReadFramesInMultipleCallsToReadFrom_whenLastPieceIsSmall', function() {
+        const frame = createFrameWithRandomBytes(1000);
+
+        const message = ClientMessage.createForEncode();
+        message.addFrame(frame);
+
+        const buffer = writeToBuffer(message);
+
+        const firstPartition = buffer.slice(0, 750);
+        const secondPartition = buffer.slice(750, 1002);
+        const thirdPartition = buffer.slice(1002);
+
+        reader.append(firstPartition);
+        reader.read();
+        reader.append(secondPartition);
+        reader.read();
+        reader.append(thirdPartition);
+
+        const messageRead = reader.read();
+        expect(messageRead.hasNextFrame()).to.be.true;
+
+        const frameRead = messageRead.nextFrame();
+        expect(frameRead.content).to.deep.equal(frame.content);
+
+        expect(messageRead.hasNextFrame()).to.be.false;
+    });
+
+    it('testRead_whenTheFrameLengthAndFlagsNotReceivedAtFirst', function() {
+        const frame = createFrameWithRandomBytes(100);
+
+        const message = ClientMessage.createForEncode();
+        message.addFrame(frame);
+
+        const buffer = writeToBuffer(message);
+        const firstPartition = buffer.slice(0, 4);
+        reader.append(firstPartition);
+        let messageRead = reader.read();
+
+        expect(messageRead).to.be.null;
+        const secondPartition = buffer.slice(4, buffer.length);
+        reader.append(secondPartition);
+
+        messageRead = reader.read();
+        expect(messageRead).to.be.not.equal(null);
+        expect(messageRead.hasNextFrame()).to.be.true;
+
+        const frameRead = messageRead.nextFrame();
+        expect(frameRead.content).to.deep.equal(frame.content);
+
+        expect(messageRead.hasNextFrame()).to.be.false;
     });
 });
